@@ -143,7 +143,7 @@ namespace OtripleS.Web.Api.Services.Foundations.SemesterCourses
 
         private static void ValidateStorageSemesterCourse(SemesterCourse storageSemesterCourse, Guid semesterCourseId)
         {
-            if (storageSemesterCourse == null)
+            if (storageSemesterCourse is null)
             {
                 throw new NotFoundSemesterCourseException(semesterCourseId);
             }
@@ -162,13 +162,15 @@ namespace OtripleS.Web.Api.Services.Foundations.SemesterCourses
         private static bool IsInvalid(DateTimeOffset input) => input == default;
         private static bool IsInvalid(Guid input) => input == default;
 
-        private bool IsDateNotRecent(DateTimeOffset dateTime)
+        private bool IsDateNotRecent(DateTimeOffset date)
         {
-            DateTimeOffset now = this.dateTimeBroker.GetCurrentDateTime();
-            int oneMinute = 1;
-            TimeSpan difference = now.Subtract(dateTime);
+            DateTimeOffset currentDateTime =
+                this.dateTimeBroker.GetCurrentDateTime();
 
-            return Math.Abs(difference.TotalMinutes) > oneMinute;
+            TimeSpan timeDifference = currentDateTime.Subtract(date);
+            TimeSpan oneMinute = TimeSpan.FromMinutes(1);
+
+            return timeDifference.Duration() > oneMinute;
         }
 
         private void ValidateAuditFieldsDataOnModify(SemesterCourse semesterCourse)
