@@ -16,7 +16,7 @@ namespace OtripleS.Web.Api.Services.Foundations.StudentExamFees
           Guid studentId,
           Guid examFeeId)
         {
-            if (storageStudentExamFee == null)
+            if (storageStudentExamFee is null)
                 throw new NotFoundStudentExamFeeException(studentId, examFeeId);
         }
 
@@ -146,13 +146,15 @@ namespace OtripleS.Web.Api.Services.Foundations.StudentExamFees
         private static bool IsInvalid(Guid input) => input == default;
         private static bool IsInvalid(DateTimeOffset input) => input == default;
 
-        private bool IsDateNotRecent(DateTimeOffset dateTime)
+        private bool IsDateNotRecent(DateTimeOffset date)
         {
-            DateTimeOffset now = this.dateTimeBroker.GetCurrentDateTime();
-            int oneMinute = 1;
-            TimeSpan difference = now.Subtract(dateTime);
+            DateTimeOffset currentDateTime =
+                this.dateTimeBroker.GetCurrentDateTime();
 
-            return Math.Abs(difference.TotalMinutes) > oneMinute;
+            TimeSpan timeDifference = currentDateTime.Subtract(date);
+            TimeSpan oneMinute = TimeSpan.FromMinutes(1);
+
+            return timeDifference.Duration() > oneMinute;
         }
     }
 }
