@@ -32,7 +32,7 @@ namespace OtripleS.Web.Api.Services.Foundations.Users
 
         private static void ValidateStorageUser(User storageUser, Guid userId)
         {
-            if (storageUser == null)
+            if (storageUser is null)
             {
                 throw new NotFoundUserException(userId);
             }
@@ -149,13 +149,15 @@ namespace OtripleS.Web.Api.Services.Foundations.Users
         private static bool IsInvalid(string input) => String.IsNullOrWhiteSpace(input);
         private static bool IsInvalid(DateTimeOffset input) => input == default;
 
-        private bool IsDateNotRecent(DateTimeOffset dateTime)
+        private bool IsDateNotRecent(DateTimeOffset date)
         {
-            DateTimeOffset now = this.dateTimeBroker.GetCurrentDateTime();
-            int oneMinute = 1;
-            TimeSpan difference = now.Subtract(dateTime);
+            DateTimeOffset currentDateTime =
+                this.dateTimeBroker.GetCurrentDateTime();
 
-            return Math.Abs(difference.TotalMinutes) >= oneMinute;
+            TimeSpan timeDifference = currentDateTime.Subtract(date);
+            TimeSpan oneMinute = TimeSpan.FromMinutes(1);
+
+            return timeDifference.Duration() > oneMinute;
         }
     }
 }
