@@ -14,36 +14,7 @@ using Xunit;
 namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.CalendarEntryAttachments
 {
     public partial class CalendarEntryAttachmentServiceTests
-    {
-        [Fact]
-        public async Task ShouldAddCalendarEntryAttachmentAsync()
-        {
-            // given
-            CalendarEntryAttachment randomCalendarEntryAttachment = CreateRandomCalendarEntryAttachment();
-            CalendarEntryAttachment inputCalendarEntryAttachment = randomCalendarEntryAttachment;
-            CalendarEntryAttachment storageCalendarEntryAttachment = randomCalendarEntryAttachment;
-            CalendarEntryAttachment expectedCalendarEntryAttachment = storageCalendarEntryAttachment;
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.InsertCalendarEntryAttachmentAsync(inputCalendarEntryAttachment))
-                    .ReturnsAsync(storageCalendarEntryAttachment);
-
-            // when
-            CalendarEntryAttachment actualCalendarEntryAttachment =
-                await this.calendarEntryAttachmentService.AddCalendarEntryAttachmentAsync(inputCalendarEntryAttachment);
-
-            // then
-            actualCalendarEntryAttachment.Should().BeEquivalentTo(expectedCalendarEntryAttachment);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.InsertCalendarEntryAttachmentAsync(inputCalendarEntryAttachment),
-                    Times.Once);
-
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-        }
-
+    {        
         [Fact]
         public void ShouldRetrieveAllCalendarEntryAttachments()
         {
