@@ -5,6 +5,8 @@
 
 using System;
 using System.Threading.Tasks;
+using EFxceptions.Models.Exceptions;
+using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using Moq;
 using OtripleS.Web.Api.Models.Exams;
@@ -16,33 +18,33 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
     public partial class ExamServiceTests
     {
         [Fact]
-        public async Task ShouldThrowDependencyExceptionOnAddWhenSqlExceptionOccursAndLogItAsync()
+        public async Task ShouldThrowCriticalDependencyExceptionOnAddIfSqlErrorOccursAndLogItAsync()
         {
             // given
-            DateTimeOffset dateTime = GetRandomDateTime();
-            Exam randomExam = CreateRandomExam(dateTime);
-            Exam inputExam = randomExam;
-            inputExam.UpdatedBy = inputExam.CreatedBy;
-            var sqlException = GetSqlException();
+            Exam someExam = CreateRandomExam();
+            SqlException sqlException = GetSqlException();
+
+            var failedExamStorageException =
+                new FailedExamStorageException(sqlException);
 
             var expectedExamDependencyException =
-                new ExamDependencyException(sqlException);
+                new ExamDependencyException(failedExamStorageException);
 
             this.dateTimeBrokerMock.Setup(broker =>
                 broker.GetCurrentDateTime())
-                    .Returns(dateTime);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.InsertExamAsync(inputExam))
-                    .ThrowsAsync(sqlException);
+                    .Throws(sqlException);
 
             // when
-            ValueTask<Exam> createExamTask =
-                this.examService.AddExamAsync(inputExam);
+            ValueTask<Exam> addExamTask =
+                this.examService.AddExamAsync(someExam);
 
             // then
             await Assert.ThrowsAsync<ExamDependencyException>(() =>
-                createExamTask.AsTask());
+                addExamTask.AsTask());
+
+            this.dateTimeBrokerMock.Verify(broker =>
+                broker.GetCurrentDateTime(),
+                    Times.Once);
 
             this.loggingBrokerMock.Verify(broker =>
                 broker.LogCritical(It.Is(SameExceptionAs(
@@ -50,12 +52,8 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
                         Times.Once);
 
             this.storageBrokerMock.Verify(broker =>
-                broker.InsertExamAsync(inputExam),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Once);
+                broker.InsertExamAsync(It.IsAny<Exam>()),
+                    Times.Never);
 
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
@@ -63,33 +61,79 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
         }
 
         [Fact]
-        public async Task ShouldThrowDependencyExceptionOnAddWhenDbExceptionOccursAndLogItAsync()
+        public async void ShouldThrowDependencyValidationExceptionOnAddIfExamAlreadyExistsAndLogItAsync()
         {
             // given
-            DateTimeOffset dateTime = GetRandomDateTime();
-            Exam randomExam = CreateRandomExam(dateTime);
-            Exam inputExam = randomExam;
-            inputExam.UpdatedBy = inputExam.CreatedBy;
-            var databaseUpdateException = new DbUpdateException();
+            Exam someExam = CreateRandomExam();
+            string exceptionMessage = GetRandomMessage();
 
-            var expectedExamDependencyException =
-                new ExamDependencyException(databaseUpdateException);
+            var duplicateKeyException =
+                new DuplicateKeyException(exceptionMessage);
+
+            var alreadyExistsExamException =
+                new AlreadyExistsExamException(duplicateKeyException);
+
+            var expectedExamDependencyValidationException =
+                new ExamDependencyValidationException(alreadyExistsExamException);
 
             this.dateTimeBrokerMock.Setup(broker =>
                 broker.GetCurrentDateTime())
-                    .Returns(dateTime);
+                    .Throws(duplicateKeyException);
 
-            this.storageBrokerMock.Setup(broker =>
-                broker.InsertExamAsync(inputExam))
-                    .ThrowsAsync(databaseUpdateException);
+            // when
+            ValueTask<Exam> addExamTask =
+                this.examService.AddExamAsync(someExam);
+
+            // then
+            await Assert.ThrowsAsync<ExamDependencyValidationException>(() =>
+                addExamTask.AsTask());
+
+            this.dateTimeBrokerMock.Verify(broker =>
+                broker.GetCurrentDateTime(),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+               broker.LogError(It.Is(SameExceptionAs(
+                   expectedExamDependencyValidationException))),
+                        Times.Once);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.InsertExamAsync(It.IsAny<Exam>()),
+                    Times.Never);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+        }
+
+        [Fact]
+        public async Task ShouldThrowDependencyExceptionOnAddIfDatabaseUpdateErrorOccursAndLogItAsync()
+        {
+            // given
+            Exam someExam = CreateRandomExam();
+            var databaseUpdateException = new DbUpdateException();
+
+            var failedExamStorageException =
+                new FailedExamStorageException(databaseUpdateException);
+
+            var expectedExamDependencyException =
+                new ExamDependencyException(failedExamStorageException);
+
+            this.dateTimeBrokerMock.Setup(broker =>
+                broker.GetCurrentDateTime())
+                    .Throws(databaseUpdateException);
 
             // when
             ValueTask<Exam> createExamTask =
-                this.examService.AddExamAsync(inputExam);
+                this.examService.AddExamAsync(someExam);
 
             // then
             await Assert.ThrowsAsync<ExamDependencyException>(() =>
                 createExamTask.AsTask());
+
+            this.dateTimeBrokerMock.Verify(broker =>
+                broker.GetCurrentDateTime(),
+                    Times.Once);
 
             this.loggingBrokerMock.Verify(broker =>
                 broker.LogError(It.Is(SameExceptionAs(
@@ -97,12 +141,8 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
                         Times.Once);
 
             this.storageBrokerMock.Verify(broker =>
-                broker.InsertExamAsync(inputExam),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Once);
+                broker.InsertExamAsync(It.IsAny<Exam>()),
+                    Times.Never);
 
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
@@ -110,13 +150,10 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
         }
 
         [Fact]
-        public async Task ShouldThrowServiceExceptionOnAddWhenExceptionOccursAndLogItAsync()
+        public async Task ShouldThrowServiceExceptionOnAddIfServiceErrorOccursAndLogItAsync()
         {
             // given
-            DateTimeOffset dateTime = GetRandomDateTime();
-            Exam randomExam = CreateRandomExam(dateTime);
-            Exam inputExam = randomExam;
-            inputExam.UpdatedBy = inputExam.CreatedBy;
+            Exam someExam = CreateRandomExam();
             var serviceException = new Exception();
 
             var failedExamServiceException =
@@ -127,19 +164,19 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
 
             this.dateTimeBrokerMock.Setup(broker =>
                 broker.GetCurrentDateTime())
-                    .Returns(dateTime);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.InsertExamAsync(inputExam))
-                    .ThrowsAsync(serviceException);
+                    .Throws(serviceException);
 
             // when
             ValueTask<Exam> createExamTask =
-                 this.examService.AddExamAsync(inputExam);
+                 this.examService.AddExamAsync(someExam);
 
             // then
             await Assert.ThrowsAsync<ExamServiceException>(() =>
                 createExamTask.AsTask());
+
+            this.dateTimeBrokerMock.Verify(broker =>
+                broker.GetCurrentDateTime(),
+                    Times.Once);
 
             this.loggingBrokerMock.Verify(broker =>
                 broker.LogError(It.Is(SameExceptionAs(
@@ -147,12 +184,8 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
                         Times.Once);
 
             this.storageBrokerMock.Verify(broker =>
-                broker.InsertExamAsync(inputExam),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Once);
+                broker.InsertExamAsync(It.IsAny<Exam>()),
+                    Times.Never);
 
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
