@@ -17,15 +17,18 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
     public partial class ExamServiceTests
     {
         [Fact]
-        public async Task ShouldThrowDependencyExceptionOnDeleteWhenSqlExceptionOccursAndLogItAsync()
+        public async Task ShouldThrowDependencyExceptionOnDeleteIfSqlErrorOccursAndLogItAsync()
         {
             // given
             Guid randomExamId = Guid.NewGuid();
             Guid inputExamId = randomExamId;
             SqlException sqlException = GetSqlException();
 
+            var failedExamStorageException =
+                new FailedExamStorageException(sqlException);
+
             var expectedExamDependencyException =
-                new ExamDependencyException(sqlException);
+                new ExamDependencyException(failedExamStorageException);
 
             this.storageBrokerMock.Setup(broker =>
                 broker.SelectExamByIdAsync(inputExamId))
@@ -54,86 +57,90 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
         }
 
         [Fact]
-        public async Task ShouldThrowDependencyExceptionOnDeleteWhenDbExceptionOccursAndLogItAsync()
+        public async Task ShouldThrowDependencyValidationExceptionOnDeleteIfDbUpdateConcurrencyErrorOccursAndLogItAsync()
         {
             // given
-            Guid randomExamId = Guid.NewGuid();
-            Guid inputExamId = randomExamId;
-            var databaseUpdateException = new DbUpdateException();
+            Guid someExamId = Guid.NewGuid();
 
-            var expectedExamDependencyException =
-                new ExamDependencyException(databaseUpdateException);
+            var databaseUpdateConcurrencyException =
+                new DbUpdateConcurrencyException();
 
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectExamByIdAsync(inputExamId))
-                    .ThrowsAsync(databaseUpdateException);
+            var lockedExamException =
+                new LockedExamException(databaseUpdateConcurrencyException);
 
-            // when
-            ValueTask<Exam> deleteExamTask =
-                this.examService.RemoveExamByIdAsync(inputExamId);
-
-            // then
-            await Assert.ThrowsAsync<ExamDependencyException>(() =>
-                deleteExamTask.AsTask());
-
-            this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameExceptionAs(
-                    expectedExamDependencyException))),
-                        Times.Once);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectExamByIdAsync(inputExamId),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.storageBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public async Task ShouldThrowDependencyExceptionOnDeleteWhenDbUpdateConcurrencyExceptionOccursAndLogItAsync()
-        {
-            // given
-            Guid randomExamId = Guid.NewGuid();
-            Guid inputExamId = randomExamId;
-            var databaseUpdateConcurrencyException = new DbUpdateConcurrencyException();
-            var lockedExamException = new LockedExamException(databaseUpdateConcurrencyException);
-
-            var expectedExamDependencyException =
-                new ExamDependencyException(lockedExamException);
+            var expectedExamDependencyValidationException =
+                new ExamDependencyValidationException(lockedExamException);
 
             this.storageBrokerMock.Setup(broker =>
-                broker.SelectExamByIdAsync(inputExamId))
+                broker.SelectExamByIdAsync(someExamId))
                     .ThrowsAsync(databaseUpdateConcurrencyException);
 
             // when
             ValueTask<Exam> deleteExamTask =
-                this.examService.RemoveExamByIdAsync(inputExamId);
+                this.examService.RemoveExamByIdAsync(someExamId);
+
+            // then
+            await Assert.ThrowsAsync<ExamDependencyValidationException>(() =>
+                deleteExamTask.AsTask());
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectExamByIdAsync(someExamId),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogError(It.Is(SameExceptionAs(
+                    expectedExamDependencyValidationException))),
+                        Times.Once);
+
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+        }
+
+        [Fact]
+        public async Task ShouldThrowDependencyExceptionOnDeleteifDatabaseUpdateErrorOccursAndLogItAsync()
+        {
+            // given
+            Guid someExamId = Guid.NewGuid();
+            var databaseUpdateException = new DbUpdateException();
+
+            var failedExamStorageException =
+                new FailedExamStorageException(databaseUpdateException);
+
+            var expectedExamDependencyException =
+                new ExamDependencyException(failedExamStorageException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectExamByIdAsync(someExamId))
+                    .ThrowsAsync(databaseUpdateException);
+
+            // when
+            ValueTask<Exam> deleteExamTask =
+                this.examService.RemoveExamByIdAsync(someExamId);
 
             // then
             await Assert.ThrowsAsync<ExamDependencyException>(() =>
                 deleteExamTask.AsTask());
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectExamByIdAsync(someExamId),
+                    Times.Once);
 
             this.loggingBrokerMock.Verify(broker =>
                 broker.LogError(It.Is(SameExceptionAs(
                     expectedExamDependencyException))),
                         Times.Once);
 
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectExamByIdAsync(inputExamId),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
             this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
         }
 
         [Fact]
-        public async Task ShouldThrowServiceExceptionOnDeleteWhenExceptionOccursAndLogItAsync()
+        public async Task ShouldThrowServiceExceptionOnDeleteIfServiceErrorOccursAndLogItAsync()
         {
             // given
-            Guid randomExamId = Guid.NewGuid();
-            Guid inputExamId = randomExamId;
+            Guid someExamId = Guid.NewGuid();
             var serviceException = new Exception();
 
             var failedExamServiceException =
@@ -143,29 +150,29 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
                 new ExamServiceException(failedExamServiceException);
 
             this.storageBrokerMock.Setup(broker =>
-                broker.SelectExamByIdAsync(inputExamId))
+                broker.SelectExamByIdAsync(someExamId))
                     .ThrowsAsync(serviceException);
 
             // when
             ValueTask<Exam> deleteExamTask =
-                this.examService.RemoveExamByIdAsync(inputExamId);
+                this.examService.RemoveExamByIdAsync(someExamId);
 
             // then
             await Assert.ThrowsAsync<ExamServiceException>(() =>
                 deleteExamTask.AsTask());
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectExamByIdAsync(someExamId),
+                    Times.Once);
 
             this.loggingBrokerMock.Verify(broker =>
                 broker.LogError(It.Is(SameExceptionAs(
                     expectedExamServiceException))),
                         Times.Once);
 
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectExamByIdAsync(inputExamId),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
             this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
         }
     }
 }
