@@ -15,11 +15,10 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
     public partial class ExamServiceTests
     {
         [Fact]
-        public async void ShouldThrowValidationExceptionOnRetrieveByIdWhenIdIsInvalidAndLogItAsync()
+        public async Task ShouldThrowValidationExceptionOnRetrieveByIdIfIdIsInvalidAndLogItAsync()
         {
             // given
-            Guid randomExamId = default;
-            Guid inputExamId = randomExamId;
+            Guid invalidExamId = Guid.Empty;
 
             var invalidExamInputException = new InvalidExamException();
 
@@ -32,14 +31,14 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
 
             // when
             ValueTask<Exam> retrieveExamByIdTask =
-                this.examService.RetrieveExamByIdAsync(inputExamId);
+                this.examService.RetrieveExamByIdAsync(invalidExamId);
 
             // then
             await Assert.ThrowsAsync<ExamValidationException>(() =>
                 retrieveExamByIdTask.AsTask());
 
             this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameValidationExceptionAs(
+                broker.LogError(It.Is(SameExceptionAs(
                     expectedExamValidationException))),
                         Times.Once);
 
@@ -57,28 +56,33 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
         }
 
         [Fact]
-        public async void ShouldThrowValidationExceptionOnRetrieveByIdWhenStorageExamIsNullAndLogItAsync()
+        public async void ShouldThrowValidationExceptionOnRetrieveByIdIfStorageExamIsNullAndLogItAsync()
         {
             // given
-            Guid randomExamId = Guid.NewGuid();
-            Guid inputExamId = randomExamId;
+            Guid someExamId = Guid.NewGuid();
             Exam invalidStorageExam = null;
-            var notFoundExamException = new NotFoundExamException(inputExamId);
+
+            var notFoundExamException =
+                new NotFoundExamException(someExamId);
 
             var expectedExamValidationException =
                 new ExamValidationException(notFoundExamException);
 
             this.storageBrokerMock.Setup(broker =>
-                broker.SelectExamByIdAsync(inputExamId))
+                broker.SelectExamByIdAsync(someExamId))
                     .ReturnsAsync(invalidStorageExam);
 
             // when
             ValueTask<Exam> retrieveExamByIdTask =
-                this.examService.RetrieveExamByIdAsync(inputExamId);
+                this.examService.RetrieveExamByIdAsync(someExamId);
 
             // then
             await Assert.ThrowsAsync<ExamValidationException>(() =>
                 retrieveExamByIdTask.AsTask());
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectExamByIdAsync(someExamId),
+                    Times.Once);
 
             this.loggingBrokerMock.Verify(broker =>
                 broker.LogError(It.Is(SameExceptionAs(
@@ -89,13 +93,9 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
                 broker.GetCurrentDateTime(),
                     Times.Never);
 
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectExamByIdAsync(inputExamId),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
             this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
         }
     }
 }
