@@ -64,22 +64,18 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Exams
         private static Exam CreateRandomExam(DateTimeOffset dateTime) =>
             CreateRandomExamFiller(dateTime).Create();
 
+        private static Exam CreateRandomExam() =>
+            CreateRandomExamFiller(dateTime: GetRandomDateTime()).Create();
+
         private static SqlException GetSqlException() =>
             (SqlException)FormatterServices.GetUninitializedObject(typeof(SqlException));
 
-        private static Expression<Func<Exception, bool>> SameExceptionAs(Exception expectedException)
+        private static Expression<Func<Xeption, bool>> SameExceptionAs(Xeption expectedException)
         {
             return actualException =>
                  actualException.Message == expectedException.Message
-                 && actualException.InnerException.Message == expectedException.InnerException.Message;
-        }
-
-        private static Expression<Func<Exception, bool>> SameValidationExceptionAs(Exception expectedException)
-        {
-            return actualException =>
-                expectedException.Message == actualException.Message &&
-                expectedException.InnerException.Message == actualException.InnerException.Message &&
-                (actualException.InnerException as Xeption).DataEquals(expectedException.InnerException.Data);
+                 && actualException.InnerException.Message == expectedException.InnerException.Message
+                 && (actualException.InnerException as Xeption).DataEquals(expectedException.InnerException.Data);
         }
 
         private static ExamType GetInValidExamType()
