@@ -18,19 +18,18 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Registrations
         public async void ShouldThrowValidationExceptionOnRetrieveByIdWhenIdIsInvalidAndLogItAsync()
         {
             //given
-            Guid randomRegistrationId = default;
-            Guid inputRegistrationId = randomRegistrationId;
+            Guid invalidRegistrationId = Guid.Empty;
 
             var invalidRegistrationException = new InvalidRegistrationException(
                 parameterName: nameof(Registration.Id),
-                parameterValue: inputRegistrationId);
+                parameterValue: invalidRegistrationId);
 
             var expectedRegistrationValidationException =
                 new RegistrationValidationException(invalidRegistrationException);
 
             //when
             ValueTask<Registration> retrieveRegistrationByIdTask =
-                this.registrationService.RetrieveRegistrationByIdAsync(inputRegistrationId);
+                this.registrationService.RetrieveRegistrationByIdAsync(invalidRegistrationId);
 
             //then
             await Assert.ThrowsAsync<RegistrationValidationException>(() => retrieveRegistrationByIdTask.AsTask());
@@ -57,10 +56,9 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Registrations
         public async void ShouldThrowValidationExceptionOnRetrieveByIdWhenStorageRegistrationIsNullAndLogItAsync()
         {
             //given
-            Guid randomRegistrationId = Guid.NewGuid();
-            Guid someRegistrationId = randomRegistrationId;
+            Guid invalidRegistrationId = Guid.NewGuid();
             Registration invalidStorageRegistration = null;
-            var notFoundRegistrationException = new NotFoundRegistrationException(someRegistrationId);
+            var notFoundRegistrationException = new NotFoundRegistrationException(invalidRegistrationId);
 
             var expectedRegistrationValidationException =
                 new RegistrationValidationException(notFoundRegistrationException);
@@ -71,7 +69,7 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Registrations
 
             //when
             ValueTask<Registration> retrieveRegistrationByIdTask =
-                this.registrationService.RetrieveRegistrationByIdAsync(someRegistrationId);
+                this.registrationService.RetrieveRegistrationByIdAsync(invalidRegistrationId);
 
             //then
             await Assert.ThrowsAsync<RegistrationValidationException>(() =>
