@@ -20,36 +20,31 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.TeacherContacts
         public async Task ShouldThrowDependencyExceptionOnRetrieveWhenSqlExceptionOccursAndLogItAsync()
         {
             // given
-            var randomContactId = Guid.NewGuid();
-            var inputContactId = randomContactId;
-            Guid randomTeacherId = Guid.NewGuid();
-            Guid inputTeacherId = randomTeacherId;
+            Guid someContactId = Guid.NewGuid();
+            Guid someTeacherId = Guid.NewGuid();
             SqlException sqlException = GetSqlException();
 
             var expectedTeacherContactDependencyException =
                 new TeacherContactDependencyException(sqlException);
 
             this.storageBrokerMock.Setup(broker =>
-                 broker.SelectTeacherContactByIdAsync(inputTeacherId, inputContactId))
+                 broker.SelectTeacherContactByIdAsync(someTeacherId, someContactId))
                     .ThrowsAsync(sqlException);
 
             // when
             ValueTask<TeacherContact> retrieveTeacherContactTask =
-                this.teacherContactService.RetrieveTeacherContactByIdAsync(
-                    inputTeacherId,
-                    inputContactId);
+                this.teacherContactService.RetrieveTeacherContactByIdAsync(someTeacherId, someContactId);
 
             // then
             await Assert.ThrowsAsync<TeacherContactDependencyException>(() =>
                 retrieveTeacherContactTask.AsTask());
 
             this.loggingBrokerMock.Verify(broker =>
-                broker.LogCritical(It.Is(SameExceptionAs(
-                    expectedTeacherContactDependencyException))),
+                broker.LogCritical(It.Is(SameExceptionAs(expectedTeacherContactDependencyException))),
                         Times.Once);
 
             this.storageBrokerMock.Verify(broker =>
-                broker.SelectTeacherContactByIdAsync(inputTeacherId, inputContactId),
+                broker.SelectTeacherContactByIdAsync(someTeacherId, someContactId),
                     Times.Once);
 
             this.loggingBrokerMock.VerifyNoOtherCalls();
@@ -60,35 +55,31 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.TeacherContacts
         public async Task ShouldThrowDependencyExceptionOnRetrieveWhenDbUpdateExceptionOccursAndLogItAsync()
         {
             // given
-            var randomContactId = Guid.NewGuid();
-            var randomTeacherId = Guid.NewGuid();
-            Guid inputContactId = randomContactId;
-            Guid inputTeacherId = randomTeacherId;
+            Guid someContactId = Guid.NewGuid();
+            Guid someTeacherId = Guid.NewGuid();
             var databaseUpdateException = new DbUpdateException();
 
             var expectedTeacherContactDependencyException =
                 new TeacherContactDependencyException(databaseUpdateException);
 
             this.storageBrokerMock.Setup(broker =>
-                broker.SelectTeacherContactByIdAsync(inputTeacherId, inputContactId))
+                broker.SelectTeacherContactByIdAsync(someTeacherId, someContactId))
                     .ThrowsAsync(databaseUpdateException);
 
             // when
             ValueTask<TeacherContact> retrieveTeacherContactTask =
-                this.teacherContactService.RetrieveTeacherContactByIdAsync
-                (inputTeacherId, inputContactId);
+                this.teacherContactService.RetrieveTeacherContactByIdAsync(someTeacherId, someContactId);
 
             // then
             await Assert.ThrowsAsync<TeacherContactDependencyException>(
                 () => retrieveTeacherContactTask.AsTask());
 
             this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameExceptionAs(
-                    expectedTeacherContactDependencyException))),
+                broker.LogError(It.Is(SameExceptionAs(expectedTeacherContactDependencyException))),
                         Times.Once);
 
             this.storageBrokerMock.Verify(broker =>
-                broker.SelectTeacherContactByIdAsync(inputTeacherId, inputContactId),
+                broker.SelectTeacherContactByIdAsync(someTeacherId, someContactId),
                     Times.Once);
 
             this.loggingBrokerMock.VerifyNoOtherCalls();
@@ -99,40 +90,34 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.TeacherContacts
         public async Task ShouldThrowServiceExceptionOnRetrieveWhenExceptionOccursAndLogItAsync()
         {
             // given
-            var randomContactId = Guid.NewGuid();
-            var randomTeacherId = Guid.NewGuid();
-            Guid inputContactId = randomContactId;
-            Guid inputTeacherId = randomTeacherId;
+            Guid someContactId = Guid.NewGuid();
+            Guid someTeacherId = Guid.NewGuid();
             var serviceException = new Exception();
 
             var failedTeacherContactException =
                 new FailedTeacherContactServiceException(serviceException);
 
             var expectedTeacherContactException =
-                new TeacherContactServiceException(
-                    failedTeacherContactException);
+                new TeacherContactServiceException(failedTeacherContactException);
 
             this.storageBrokerMock.Setup(broker =>
-                broker.SelectTeacherContactByIdAsync(inputTeacherId, inputContactId))
+                broker.SelectTeacherContactByIdAsync(someTeacherId, someContactId))
                     .ThrowsAsync(serviceException);
 
             // when
             ValueTask<TeacherContact> retrieveTeacherContactTask =
-                this.teacherContactService.RetrieveTeacherContactByIdAsync(
-                    inputTeacherId,
-                    inputContactId);
+                this.teacherContactService.RetrieveTeacherContactByIdAsync(someTeacherId, someContactId);
 
             // then
             await Assert.ThrowsAsync<TeacherContactServiceException>(() =>
                 retrieveTeacherContactTask.AsTask());
 
             this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameExceptionAs(
-                    expectedTeacherContactException))),
+                broker.LogError(It.Is(SameExceptionAs(expectedTeacherContactException))),
                         Times.Once);
 
             this.storageBrokerMock.Verify(broker =>
-                broker.SelectTeacherContactByIdAsync(inputTeacherId, inputContactId),
+                broker.SelectTeacherContactByIdAsync(someTeacherId, someContactId),
                     Times.Once);
 
             this.loggingBrokerMock.VerifyNoOtherCalls();
