@@ -15,45 +15,7 @@ using Xunit;
 namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Teachers
 {
     public partial class TeacherServiceTests
-    {
-        [Fact]
-        public async Task ShouldDeleteTeacherAsync()
-        {
-            // given
-            DateTimeOffset dateTime = GetRandomDateTime();
-            Teacher randomTeacher = CreateRandomTeacher(dateTime);
-            Guid inputTeacherId = randomTeacher.Id;
-            Teacher inputTeacher = randomTeacher;
-            Teacher storageTeacher = randomTeacher;
-            Teacher expectedTeacher = randomTeacher;
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectTeacherByIdAsync(inputTeacherId))
-                    .ReturnsAsync(inputTeacher);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.DeleteTeacherAsync(inputTeacher))
-                    .ReturnsAsync(storageTeacher);
-
-            // when
-            Teacher actualTeacher =
-                await this.teacherService.RemoveTeacherByIdAsync(inputTeacherId);
-
-            // then
-            actualTeacher.Should().BeEquivalentTo(expectedTeacher);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectTeacherByIdAsync(inputTeacherId),
-                    Times.Once);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.DeleteTeacherAsync(inputTeacher),
-                    Times.Once);
-
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-        }
+    {        
 
         [Fact]
         public async Task ShouldRetrieveTeacherByIdAsync()
@@ -83,39 +45,7 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Teachers
             this.storageBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public void ShouldRetrieveAllTeachers()
-        {
-            // given
-            IQueryable<Teacher> randomTeachers = CreateRandomTeachers();
-            IQueryable<Teacher> storageTeachers = randomTeachers;
-            IQueryable<Teacher> expectedTeachers = storageTeachers;
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectAllTeachers())
-                    .Returns(storageTeachers);
-
-            // when
-            IQueryable<Teacher> actualTeachers =
-                this.teacherService.RetrieveAllTeachers();
-
-            // then
-            actualTeachers.Should().BeEquivalentTo(expectedTeachers);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Never);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectAllTeachers(),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-        }
+        }        
 
         [Fact]
         public async Task ShouldCreateTeacherAsync()
@@ -155,58 +85,6 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Teachers
             this.storageBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public async Task ShouldModifyTeacherAsync()
-        {
-            // given
-            int randomNumber = GetRandomNumber();
-            int randomDays = randomNumber;
-            DateTimeOffset randomDate = GetRandomDateTime();
-            DateTimeOffset randomInputDate = GetRandomDateTime();
-            Teacher randomTeacher = CreateRandomTeacher(randomInputDate);
-            Teacher inputTeacher = randomTeacher;
-            Teacher afterUpdateStorageTeacher = inputTeacher;
-            Teacher expectedTeacher = afterUpdateStorageTeacher;
-            Teacher beforeUpdateStorageTeacher = randomTeacher.DeepClone();
-            inputTeacher.UpdatedDate = randomDate;
-            Guid studentId = inputTeacher.Id;
-
-            this.dateTimeBrokerMock.Setup(broker =>
-               broker.GetCurrentDateTime())
-                   .Returns(randomDate);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectTeacherByIdAsync(studentId))
-                    .ReturnsAsync(beforeUpdateStorageTeacher);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.UpdateTeacherAsync(inputTeacher))
-                    .ReturnsAsync(afterUpdateStorageTeacher);
-
-            // when
-            Teacher actualTeacher =
-                await this.teacherService.ModifyTeacherAsync(inputTeacher);
-
-            // then
-            actualTeacher.Should().BeEquivalentTo(expectedTeacher);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Once);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectTeacherByIdAsync(studentId),
-                    Times.Once);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.UpdateTeacherAsync(inputTeacher),
-                    Times.Once);
-
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-        }
+        }        
     }
 }
