@@ -4,10 +4,11 @@
 // ---------------------------------------------------------------
 
 using System;
+using Xeptions;
 
 namespace OtripleS.Web.Api.Models.Exams.Exceptions
 {
-    public class AlreadyExistsExamException : Exception
+    public class AlreadyExistsExamException : Xeption
     {
         public AlreadyExistsExamException(Exception innerException)
             : base(message: "Exam with the same id already exists.", innerException) { }
