@@ -3,13 +3,13 @@
 // FREE TO USE AS LONG AS SOFTWARE FUNDS ARE DONATED TO THE POOR
 // ---------------------------------------------------------------
 
-using System;
+using Xeptions;
 
 namespace OtripleS.Web.Api.Models.Exams.Exceptions
 {
-    public class ExamDependencyException : Exception
+    public class ExamDependencyException : Xeption
     {
-        public ExamDependencyException(Exception innerException)
-            : base(message: "Service dependency error occurred, contact support.", innerException) { }
+        public ExamDependencyException(Xeption innerException)
+             : base(message: "Exam dependency error occurred, contact support.", innerException) { }
     }
 }
