@@ -3,13 +3,13 @@
 // FREE TO USE AS LONG AS SOFTWARE FUNDS ARE DONATED TO THE POOR
 // ---------------------------------------------------------------
 
-using System;
+using Xeptions;
 
 namespace OtripleS.Web.Api.Models.Exams.Exceptions
 {
-    public class ExamServiceException : Exception
+    public class ExamServiceException : Xeption
     {
-        public ExamServiceException(Exception innerException)
-            : base(message: "Service error occurred, contact support.", innerException) { }
+        public ExamServiceException(Xeption innerException)
+            : base(message: "Exam service error occurred, contact support.", innerException) { }
     }
 }
