@@ -11,7 +11,7 @@ namespace OtripleS.Web.Api.Models.Exams.Exceptions
     public class FailedExamServiceException : Xeption
     {
         public FailedExamServiceException(Exception innerException)
-            : base(message: "Failed exam service error occured, contact support",
+            : base(message: "Failed exam service error occurred, contact support",
                   innerException)
         { }
     }
