@@ -165,13 +165,15 @@ namespace OtripleS.Web.Api.Services.Foundations.Attendances
         private static bool IsInvalid(Guid inputId) => inputId == default;
         private static bool IsInvalid(DateTimeOffset inputDateTimeOffset) => inputDateTimeOffset == default;
 
-        private bool IsDateNotRecent(DateTimeOffset dateTime)
+        private bool IsDateNotRecent(DateTimeOffset date)
         {
-            DateTimeOffset now = this.dateTimeBroker.GetCurrentDateTime();
-            int oneMinute = 1;
-            TimeSpan difference = now.Subtract(dateTime);
+            DateTimeOffset currentDateTime =
+                this.dateTimeBroker.GetCurrentDateTime();
 
-            return Math.Abs(difference.TotalMinutes) > oneMinute;
+            TimeSpan timeDifference = currentDateTime.Subtract(date);
+            TimeSpan oneMinute = TimeSpan.FromMinutes(1);
+
+            return timeDifference.Duration() > oneMinute;
         }
 
         private static void Validate(params (dynamic Rule, string Parameter)[] validations)
