@@ -73,6 +73,7 @@ namespace OtripleS.Web.Api.Services.Foundations.StudentGuardians
                     parameterName: nameof(StudentGuardian.StudentId),
                     parameterValue: studentId);
             }
+            
             if (guardianId == default)
             {
                 throw new InvalidStudentGuardianInputException(
@@ -179,13 +180,15 @@ namespace OtripleS.Web.Api.Services.Foundations.StudentGuardians
         private static bool IsInvalid(DateTimeOffset input) => input == default;
         private static bool IsInvalid(Guid input) => input == default;
 
-        private bool IsDateNotRecent(DateTimeOffset dateTime)
+        private bool IsDateNotRecent(DateTimeOffset date)
         {
-            DateTimeOffset now = this.dateTimeBroker.GetCurrentDateTime();
-            int oneMinute = 1;
-            TimeSpan difference = now.Subtract(dateTime);
+            DateTimeOffset currentDateTime =
+                this.dateTimeBroker.GetCurrentDateTime();
 
-            return Math.Abs(difference.TotalMinutes) > oneMinute;
+            TimeSpan timeDifference = currentDateTime.Subtract(date);
+            TimeSpan oneMinute = TimeSpan.FromMinutes(1);
+
+            return timeDifference.Duration() > oneMinute;
         }
     }
 }
